refactor: modularise todos feature
